Move attibute queries to attribute graphql module
