+++ saleor/graphql/api.py
@@ -2,7 +2,7 @@ from graphene_federation import build_schema
 
 from .account.schema import AccountMutations, AccountQueries
 from .app.schema import AppMutations, AppQueries
-from .attribute.schema import AttributeMutations
+from .attribute.schema import AttributeMutations, AttributeQueries
 from .checkout.schema import CheckoutMutations, CheckoutQueries
 from .core.schema import CoreQueries
 from .csv.schema import CsvMutations, CsvQueries
@@ -26,6 +26,7 @@ from .webhook.schema import WebhookMutations, WebhookQueries
 class Query(
     AccountQueries,
     AppQueries,
+    AttributeQueries,
     CheckoutQueries,
     CoreQueries,
     CsvQueries,
