+++ saleor/graphql/csv/mutations.py
@@ -7,11 +7,12 @@ from ...core.permissions import ProductPermissions
 from ...csv import models as csv_models
 from ...csv.events import export_started_event
 from ...csv.tasks import export_products_task
+from ..attribute.types import Attribute
 from ..core.enums import ExportErrorCode
 from ..core.mutations import BaseMutation
 from ..core.types.common import ExportError
 from ..product.filters import ProductFilterInput
-from ..product.types import Attribute, Product
+from ..product.types import Product
 from ..utils import resolve_global_ids_to_primary_keys
 from ..warehouse.types import Warehouse
 from .enums import ExportScope, FileTypeEnum, ProductFieldEnum
