+++ saleor/graphql/page/dataloaders.py
@@ -5,11 +5,8 @@ from promise import Promise
 from ...attribute.models import AssignedPageAttribute, AttributePage, AttributeValue
 from ...core.permissions import PagePermissions
 from ...page.models import Page, PageType
+from ..attribute.dataloaders import AttributesByAttributeId, AttributeValueByIdLoader
 from ..core.dataloaders import DataLoader
-from ..product.dataloaders.attributes import (
-    AttributesByAttributeId,
-    AttributeValueByIdLoader,
-)
 
 
 class PageByIdLoader(DataLoader):
