+++ saleor/graphql/product/dataloaders/__init__.py
@@ -1,5 +1,4 @@
 from .attributes import (
-    AttributeValuesByAttributeIdLoader,
     ProductAttributesByProductTypeIdLoader,
     SelectedAttributesByProductIdLoader,
     SelectedAttributesByProductVariantIdLoader,
@@ -18,7 +17,6 @@ from .products import (
 )
 
 __all__ = [
-    "AttributeValuesByAttributeIdLoader",
     "CategoryByIdLoader",
     "CollectionByIdLoader",
     "CollectionsByProductIdLoader",
