+++ saleor/graphql/product/sorters.py
@@ -6,57 +6,6 @@ from ...product.models import Category, Product
 from ..core.types import SortInputObjectType
 
 
-class AttributeSortField(graphene.Enum):
-    NAME = ["name", "slug"]
-    SLUG = ["slug"]
-    VALUE_REQUIRED = ["value_required", "name", "slug"]
-    IS_VARIANT_ONLY = ["is_variant_only", "name", "slug"]
-    VISIBLE_IN_STOREFRONT = ["visible_in_storefront", "name", "slug"]
-    FILTERABLE_IN_STOREFRONT = ["filterable_in_storefront", "name", "slug"]
-    FILTERABLE_IN_DASHBOARD = ["filterable_in_dashboard", "name", "slug"]
-    STOREFRONT_SEARCH_POSITION = ["storefront_search_position", "name", "pk"]
-    AVAILABLE_IN_GRID = ["available_in_grid", "name", "pk"]
-
-    @property
-    def description(self):
-        # pylint: disable=no-member
-        descriptions = {
-            AttributeSortField.NAME.name: "Sort attributes by name",
-            AttributeSortField.SLUG.name: "Sort attributes by slug",
-            AttributeSortField.VALUE_REQUIRED.name: (
-                "Sort attributes by the value required flag"
-            ),
-            AttributeSortField.IS_VARIANT_ONLY.name: (
-                "Sort attributes by the variant only flag"
-            ),
-            AttributeSortField.VISIBLE_IN_STOREFRONT.name: (
-                "Sort attributes by visibility in the storefront"
-            ),
-            AttributeSortField.FILTERABLE_IN_STOREFRONT.name: (
-                "Sort attributes by the filterable in storefront flag"
-            ),
-            AttributeSortField.FILTERABLE_IN_DASHBOARD.name: (
-                "Sort attributes by the filterable in dashboard flag"
-            ),
-            AttributeSortField.STOREFRONT_SEARCH_POSITION.name: (
-                "Sort attributes by their position in storefront"
-            ),
-            AttributeSortField.AVAILABLE_IN_GRID.name: (
-                "Sort attributes based on whether they can be displayed "
-                "or not in a product grid."
-            ),
-        }
-        if self.name in descriptions:
-            return descriptions[self.name]
-        raise ValueError("Unsupported enum value: %s" % self.value)
-
-
-class AttributeSortingInput(SortInputObjectType):
-    class Meta:
-        sort_enum = AttributeSortField
-        type_name = "attributes"
-
-
 class CategorySortField(graphene.Enum):
     NAME = ["name", "slug"]
     PRODUCT_COUNT = ["product_count", "name", "slug"]
