+++ saleor/graphql/product/tests/test_attributes.py
@@ -1,10 +1,7 @@
-from typing import Union
 from unittest import mock
 
 import graphene
 import pytest
-from django.db.models import Q
-from graphene.utils.str_converters import to_camel_case
 
 from ....attribute import AttributeInputType, AttributeType
 from ....attribute.models import (
@@ -13,141 +10,12 @@ from ....attribute.models import (
     AttributeValue,
     AttributeVariant,
 )
-from ....attribute.utils import associate_attribute_values_to_instance
 from ....product.error_codes import ProductErrorCode
-from ....product.models import (
-    Category,
-    Collection,
-    Product,
-    ProductType,
-    ProductVariant,
-)
-from ...attribute.enums import AttributeTypeEnum, AttributeValueType
+from ....product.models import ProductType
+from ...attribute.enums import AttributeTypeEnum
 from ...core.utils import snake_to_camel_case
 from ...tests.utils import get_graphql_content
 from ..enums import ProductAttributeType
-from ..filters import filter_attributes_by_product_types
-from ..types.attributes import resolve_attribute_value_type
-
-
-def test_get_single_attribute_by_pk(user_api_client, color_attribute_without_values):
-    attribute_gql_id = graphene.Node.to_global_id(
-        "Attribute", color_attribute_without_values.id
-    )
-    query = """
-    query($id: ID!) {
-        attribute(id: $id) {
-            id
-            slug
-        }
-    }
-    """
-    content = get_graphql_content(
-        user_api_client.post_graphql(query, {"id": attribute_gql_id})
-    )
-
-    assert content["data"]["attribute"], "Should have found an attribute"
-    assert content["data"]["attribute"]["id"] == attribute_gql_id
-    assert content["data"]["attribute"]["slug"] == color_attribute_without_values.slug
-
-
-QUERY_ATTRIBUTES = """
-    query {
-        attributes(first: 20) {
-            edges {
-                node {
-                    id
-                    name
-                    slug
-                    values {
-                        id
-                        name
-                        slug
-                    }
-                }
-            }
-        }
-    }
-"""
-
-
-def test_attributes_query(user_api_client, product):
-    attributes = Attribute.objects
-    query = QUERY_ATTRIBUTES
-    response = user_api_client.post_graphql(query)
-    content = get_graphql_content(response)
-    attributes_data = content["data"]["attributes"]["edges"]
-    assert attributes_data
-    assert len(attributes_data) == attributes.count()
-
-
-NOT_EXISTS_IDS_ATTRIBUTES_QUERY = """
-    query ($filter: AttributeFilterInput!) {
-        attributes(first: 5, filter: $filter) {
-            edges {
-                node {
-                    id
-                    name
-                }
-            }
-        }
-    }
-"""
-
-
-def test_attributes_query_ids_not_exists(user_api_client, category):
-    query = NOT_EXISTS_IDS_ATTRIBUTES_QUERY
-    variables = {"filter": {"ids": ["ygRqjpmXYqaTD9r=", "PBa4ZLBhnXHSz6v="]}}
-    response = user_api_client.post_graphql(query, variables)
-    content = get_graphql_content(response, ignore_errors=True)
-    message_error = '{"ids": [{"message": "Invalid ID specified.", "code": ""}]}'
-
-    assert len(content["errors"]) == 1
-    assert content["errors"][0]["message"] == message_error
-    assert content["data"]["attributes"] is None
-
-
-def test_attributes_query_hidden_attribute(user_api_client, product, color_attribute):
-    query = QUERY_ATTRIBUTES
-
-    # hide the attribute
-    color_attribute.visible_in_storefront = False
-    color_attribute.save(update_fields=["visible_in_storefront"])
-
-    attribute_count = Attribute.objects.get_visible_to_user(
-        user_api_client.user
-    ).count()
-    assert attribute_count == 1
-
-    response = user_api_client.post_graphql(query)
-    content = get_graphql_content(response)
-    attributes_data = content["data"]["attributes"]["edges"]
-    assert len(attributes_data) == attribute_count
-
-
-def test_attributes_query_hidden_attribute_as_staff_user(
-    staff_api_client, product, color_attribute, permission_manage_products
-):
-    query = QUERY_ATTRIBUTES
-
-    # hide the attribute
-    color_attribute.visible_in_storefront = False
-    color_attribute.save(update_fields=["visible_in_storefront"])
-
-    attribute_count = Attribute.objects.all().count()
-
-    # The user doesn't have the permission yet to manage products,
-    # the user shouldn't be able to see the hidden attributes
-    assert Attribute.objects.get_visible_to_user(staff_api_client.user).count() == 1
-
-    # The user should now be able to see the attributes
-    staff_api_client.user.user_permissions.add(permission_manage_products)
-
-    response = staff_api_client.post_graphql(query)
-    content = get_graphql_content(response)
-    attributes_data = content["data"]["attributes"]["edges"]
-    assert len(attributes_data) == attribute_count
-
 
 QUERY_PRODUCT_AND_VARIANTS_ATTRIBUTES = """
     {
@@ -325,135 +193,6 @@ def test_resolve_attribute_values_non_assigned_to_node(
 
     assert variant_attributes[0]["attribute"]["slug"] == "variant"
     assert variant_attributes[0]["values"] == []
-
-
-def test_attributes_filter_by_product_type_with_empty_value():
-    """Ensure passing an empty or null value is ignored and the queryset is simply
-    returned without any modification.
-    """
-
-    qs = Attribute.objects.all()
-
-    assert filter_attributes_by_product_types(qs, "...", "", None) is qs
-    assert filter_attributes_by_product_types(qs, "...", None, None) is qs
-
-
-def test_attributes_filter_by_product_type_with_unsupported_field(customer_user):
-    """Ensure using an unknown field to filter attributes by raises a NotImplemented
-    exception.
-    """
-
-    qs = Attribute.objects.all()
-
-    with pytest.raises(NotImplementedError) as exc:
-        filter_attributes_by_product_types(qs, "in_space", "a-value", customer_user)
-
-    assert exc.value.args == ("Filtering by in_space is unsupported",)
-
-
-def test_attributes_filter_by_non_existing_category_id(customer_user):
-    """Ensure using a non-existing category ID returns an empty query set."""
-
-    category_id = graphene.Node.to_global_id("Category", -1)
-    mocked_qs = mock.MagicMock()
-    qs = filter_attributes_by_product_types(
-        mocked_qs, "in_category", category_id, customer_user
-    )
-    assert qs == mocked_qs.none.return_value
-
-
-@pytest.mark.parametrize("tested_field", ["inCategory", "inCollection"])
-def test_attributes_in_collection_query(
-    user_api_client,
-    product_type,
-    category,
-    collection,
-    collection_with_products,
-    tested_field,
-):
-    if "Collection" in tested_field:
-        filtered_by_node_id = graphene.Node.to_global_id("Collection", collection.pk)
-    elif "Category" in tested_field:
-        filtered_by_node_id = graphene.Node.to_global_id("Category", category.pk)
-    else:
-        raise AssertionError(tested_field)
-    expected_qs = Attribute.objects.filter(
-        Q(attributeproduct__product_type_id=product_type.pk)
-        | Q(attributevariant__product_type_id=product_type.pk)
-    )
-
-    # Create another product type and attribute that shouldn't get matched
-    other_category = Category.objects.create(name="Other Category", slug="other-cat")
-    other_attribute = Attribute.objects.create(
-        name="Other", slug="other", type=AttributeType.PRODUCT_TYPE
-    )
-    other_product_type = ProductType.objects.create(
-        name="Other type", has_variants=True, is_shipping_required=True
-    )
-    other_product_type.product_attributes.add(other_attribute)
-    other_product = Product.objects.create(
-        name="Another Product",
-        product_type=other_product_type,
-        category=other_category,
-        is_published=True,
-    )
-
-    # Create another collection with products but shouldn't get matched
-    # as we don't look for this other collection
-    other_collection = Collection.objects.create(
-        name="Other Collection",
-        slug="other-collection",
-        is_published=True,
-        description="Description",
-    )
-    other_collection.products.add(other_product)
-
-    query = """
-    query($nodeID: ID!) {
-        attributes(first: 20, %(filter_input)s) {
-            edges {
-                node {
-                    id
-                    name
-                    slug
-                }
-            }
-        }
-    }
-    """
-
-    query = query % {"filter_input": "filter: { %s: $nodeID }" % tested_field}
-
-    variables = {"nodeID": filtered_by_node_id}
-    content = get_graphql_content(user_api_client.post_graphql(query, variables))
-    attributes_data = content["data"]["attributes"]["edges"]
-
-    flat_attributes_data = [attr["node"]["slug"] for attr in attributes_data]
-    expected_flat_attributes_data = list(expected_qs.values_list("slug", flat=True))
-
-    assert flat_attributes_data == expected_flat_attributes_data
-
-
-@pytest.mark.parametrize(
-    "raw_value, expected_type",
-    [
-        ("#0000", AttributeValueType.COLOR),
-        ("#FF69B4", AttributeValueType.COLOR),
-        ("rgb(255, 0, 0)", AttributeValueType.COLOR),
-        ("hsl(0, 100%, 50%)", AttributeValueType.COLOR),
-        ("hsla(120,  60%, 70%, 0.3)", AttributeValueType.COLOR),
-        ("rgba(100%, 255, 0, 0)", AttributeValueType.COLOR),
-        ("http://example.com", AttributeValueType.URL),
-        ("https://example.com", AttributeValueType.URL),
-        ("ftp://example.com", AttributeValueType.URL),
-        ("example.com", AttributeValueType.STRING),
-        ("Foo", AttributeValueType.STRING),
-        ("linear-gradient(red, yellow)", AttributeValueType.GRADIENT),
-        ("radial-gradient(#0000, yellow)", AttributeValueType.GRADIENT),
-    ],
-)
-def test_resolve_attribute_value_type(raw_value, expected_type):
-    assert resolve_attribute_value_type(raw_value) == expected_type
 
 
 def test_resolve_assigned_attribute_without_values(api_client, product_type, product):
@@ -995,51 +734,6 @@ def test_retrieve_product_attributes_input_type(
         assert gql_attr["values"][0]["inputType"] == "DROPDOWN"
 
 
-@pytest.mark.parametrize(
-    "attribute, expected_value",
-    (
-        ("filterable_in_storefront", True),
-        ("filterable_in_dashboard", True),
-        ("visible_in_storefront", True),
-        ("available_in_grid", True),
-        ("value_required", False),
-        ("storefront_search_position", 0),
-    ),
-)
-def test_retrieving_the_restricted_attributes_restricted(
-    staff_api_client,
-    color_attribute,
-    permission_manage_products,
-    attribute,
-    expected_value,
-):
-    """Checks if the attributes are restricted and if their default value
-    is the expected one."""
-
-    attribute = to_camel_case(attribute)
-    query = (
-        """
-        {
-          attributes(first: 10) {
-            edges {
-              node {
-                %s
-              }
-            }
-          }
-        }
-    """
-        % attribute
-    )
-
-    found_attributes = get_graphql_content(
-        staff_api_client.post_graphql(query, permissions=[permission_manage_products])
-    )["data"]["attributes"]["edges"]
-
-    assert len(found_attributes) == 1
-    assert found_attributes[0]["node"][attribute] == expected_value
-
-
 ATTRIBUTES_RESORT_QUERY = """
     mutation ProductTypeReorderAttributes(
       $productTypeId: ID!
@@ -1204,1068 +898,3 @@ def test_sort_attributes_within_product_type(
         gql_type, gql_attr_id = graphene.Node.from_global_id(attr["id"])
         assert gql_type == "Attribute"
         assert int(gql_attr_id) == expected_pk
-
-
-ATTRIBUTE_VALUES_RESORT_QUERY = """
-    mutation attributeReorderValues($attributeId: ID!, $moves: [ReorderInput]!) {
-      attributeReorderValues(attributeId: $attributeId, moves: $moves) {
-        attribute {
-          id
-          values {
-            id
-          }
-        }
-
-        errors {
-          field
-          message
-        }
-      }
-    }
-"""
-
-
-def test_sort_values_within_attribute_invalid_product_type(
-    staff_api_client, permission_manage_product_types_and_attributes
-):
-    """Try to reorder an invalid attribute (invalid ID)."""
-
-    attribute_id = graphene.Node.to_global_id("Attribute", -1)
-    value_id = graphene.Node.to_global_id("AttributeValue", -1)
-
-    variables = {
-        "attributeId": attribute_id,
-        "moves": [{"id": value_id, "sortOrder": 1}],
-    }
-
-    content = get_graphql_content(
-        staff_api_client.post_graphql(
-            ATTRIBUTE_VALUES_RESORT_QUERY,
-            variables,
-            permissions=[permission_manage_product_types_and_attributes],
-        )
-    )["data"]["attributeReorderValues"]
-
-    assert content["errors"] == [
-        {
-            "field": "attributeId",
-            "message": f"Couldn't resolve to an attribute: {attribute_id}",
-        }
-    ]
-
-
-def test_sort_values_within_attribute_invalid_id(
-    staff_api_client, permission_manage_product_types_and_attributes, color_attribute
-):
-    """Try to reorder a value not associated to the given attribute."""
-
-    attribute_id = graphene.Node.to_global_id("Attribute", color_attribute.id)
-    value_id = graphene.Node.to_global_id("AttributeValue", -1)
-
-    variables = {
-        "type": "VARIANT",
-        "attributeId": attribute_id,
-        "moves": [{"id": value_id, "sortOrder": 1}],
-    }
-
-    content = get_graphql_content(
-        staff_api_client.post_graphql(
-            ATTRIBUTE_VALUES_RESORT_QUERY,
-            variables,
-            permissions=[permission_manage_product_types_and_attributes],
-        )
-    )["data"]["attributeReorderValues"]
-
-    assert content["errors"] == [
-        {
-            "field": "moves",
-            "message": f"Couldn't resolve to an attribute value: {value_id}",
-        }
-    ]
-
-
-def test_sort_values_within_attribute(
-    staff_api_client, color_attribute, permission_manage_product_types_and_attributes
-):
-    attribute = color_attribute
-    AttributeValue.objects.create(attribute=attribute, name="Green", slug="green")
-    values = list(attribute.values.all())
-    assert len(values) == 3
-
-    staff_api_client.user.user_permissions.add(
-        permission_manage_product_types_and_attributes
-    )
-
-    attribute_id = graphene.Node.to_global_id("Attribute", attribute.id)
-    m2m_values = attribute.values
-    m2m_values.set(values)
-
-    assert values == sorted(
-        values, key=lambda o: o.sort_order if o.sort_order is not None else o.pk
-    ), "The values are not properly ordered"
-
-    variables = {
-        "attributeId": attribute_id,
-        "moves": [
-            {
-                "id": graphene.Node.to_global_id("AttributeValue", values[0].pk),
-                "sortOrder": +1,
-            },
-            {
-                "id": graphene.Node.to_global_id("AttributeValue", values[2].pk),
-                "sortOrder": -1,
-            },
-        ],
-    }
-
-    expected_order = [values[1].pk, values[2].pk, values[0].pk]
-
-    content = get_graphql_content(
-        staff_api_client.post_graphql(ATTRIBUTE_VALUES_RESORT_QUERY, variables)
-    )["data"]["attributeReorderValues"]
-    assert not content["errors"]
-
-    assert content["attribute"]["id"] == attribute_id
-
-    gql_values = content["attribute"]["values"]
-    assert len(gql_values) == len(expected_order)
-
-    actual_order = []
-
-    for attr, expected_pk in zip(gql_values, expected_order):
-        gql_type, gql_attr_id = graphene.Node.from_global_id(attr["id"])
-        assert gql_type == "AttributeValue"
-        actual_order.append(int(gql_attr_id))
-
-    assert actual_order == expected_order
-
-
-ATTRIBUTES_FILTER_QUERY = """
-    query($filters: AttributeFilterInput!) {
-      attributes(first: 10, filter: $filters) {
-        edges {
-          node {
-            name
-            slug
-          }
-        }
-      }
-    }
-"""
-
-
-def test_search_attributes(api_client, color_attribute, size_attribute):
-    variables = {"filters": {"search": "color"}}
-
-    attributes = get_graphql_content(
-        api_client.post_graphql(ATTRIBUTES_FILTER_QUERY, variables)
-    )["data"]["attributes"]["edges"]
-
-    assert len(attributes) == 1
-    assert attributes[0]["node"]["slug"] == "color"
-
-
-def test_filter_attributes_if_filterable_in_dashboard(
-    api_client, color_attribute, size_attribute
-):
-    color_attribute.filterable_in_dashboard = False
-    color_attribute.save(update_fields=["filterable_in_dashboard"])
-
-    variables = {"filters": {"filterableInDashboard": True}}
-
-    attributes = get_graphql_content(
-        api_client.post_graphql(ATTRIBUTES_FILTER_QUERY, variables)
-    )["data"]["attributes"]["edges"]
-
-    assert len(attributes) == 1
-    assert attributes[0]["node"]["slug"] == "size"
-
-
-def test_filter_attributes_if_available_in_grid(
-    api_client, color_attribute, size_attribute
-):
-    color_attribute.available_in_grid = False
-    color_attribute.save(update_fields=["available_in_grid"])
-
-    variables = {"filters": {"availableInGrid": True}}
-
-    attributes = get_graphql_content(
-        api_client.post_graphql(ATTRIBUTES_FILTER_QUERY, variables)
-    )["data"]["attributes"]["edges"]
-
-    assert len(attributes) == 1
-    assert attributes[0]["node"]["slug"] == "size"
-
-
-def test_filter_attributes_by_global_id_list(api_client, product_type_attribute_list):
-    global_ids = [
-        graphene.Node.to_global_id("Attribute", attribute.pk)
-        for attribute in product_type_attribute_list[:2]
-    ]
-    variables = {"filters": {"ids": global_ids}}
-
-    expected_slugs = sorted(
-        [product_type_attribute_list[0].slug, product_type_attribute_list[1].slug]
-    )
-
-    attributes = get_graphql_content(
-        api_client.post_graphql(ATTRIBUTES_FILTER_QUERY, variables)
-    )["data"]["attributes"]["edges"]
-
-    assert len(attributes) == 2
-    received_slugs = sorted(
-        [attributes[0]["node"]["slug"], attributes[1]["node"]["slug"]]
-    )
-
-    assert received_slugs == expected_slugs
-
-
-def test_filter_attributes_in_category_not_visible_in_listings_by_customer(
-    user_api_client, product_list, weight_attribute
-):
-    # given
-    product_type = ProductType.objects.create(
-        name="Default Type 2",
-        slug="default-type-2",
-        has_variants=True,
-        is_shipping_required=True,
-    )
-    product_type.product_attributes.add(weight_attribute)
-
-    last_product = product_list[-1]
-    last_product.product_type = product_type
-    last_product.visible_in_listings = False
-    last_product.save(update_fields=["visible_in_listings", "product_type"])
-
-    associate_attribute_values_to_instance(
-        product_list[-1], weight_attribute, weight_attribute.values.first()
-    )
-
-    attribute_count = Attribute.objects.count()
-
-    category = last_product.category
-    variables = {
-        "filters": {"inCategory": graphene.Node.to_global_id("Category", category.pk)}
-    }
-
-    # when
-    attributes = get_graphql_content(
-        user_api_client.post_graphql(ATTRIBUTES_FILTER_QUERY, variables)
-    )["data"]["attributes"]["edges"]
-
-    # then
-    assert len(attributes) == attribute_count - 1
-    assert weight_attribute.slug not in {
-        attribute["node"]["slug"] for attribute in attributes
-    }
-
-
-def test_filter_attributes_in_category_not_visible_in_listings_by_staff_with_perm(
-    staff_api_client, product_list, weight_attribute, permission_manage_products
-):
-    # given
-    staff_api_client.user.user_permissions.add(permission_manage_products)
-
-    product_type = ProductType.objects.create(
-        name="Default Type 2",
-        slug="default-type-2",
-        has_variants=True,
-        is_shipping_required=True,
-    )
-    product_type.product_attributes.add(weight_attribute)
-
-    last_product = product_list[-1]
-    last_product.product_type = product_type
-    last_product.visible_in_listings = False
-    last_product.save(update_fields=["visible_in_listings", "product_type"])
-
-    associate_attribute_values_to_instance(
-        product_list[-1], weight_attribute, weight_attribute.values.first()
-    )
-
-    attribute_count = Attribute.objects.count()
-
-    category = last_product.category
-    variables = {
-        "filters": {"inCategory": graphene.Node.to_global_id("Category", category.pk)}
-    }
-
-    # when
-    attributes = get_graphql_content(
-        staff_api_client.post_graphql(ATTRIBUTES_FILTER_QUERY, variables)
-    )["data"]["attributes"]["edges"]
-
-    # then
-    assert len(attributes) == attribute_count
-
-
-def test_filter_attributes_in_category_not_visible_in_listings_by_staff_without_perm(
-    staff_api_client, product_list, weight_attribute, permission_manage_products
-):
-    # given
-    product_type = ProductType.objects.create(
-        name="Default Type 2",
-        slug="default-type-2",
-        has_variants=True,
-        is_shipping_required=True,
-    )
-    product_type.product_attributes.add(weight_attribute)
-
-    last_product = product_list[-1]
-    last_product.product_type = product_type
-    last_product.visible_in_listings = False
-    last_product.save(update_fields=["visible_in_listings", "product_type"])
-
-    associate_attribute_values_to_instance(
-        product_list[-1], weight_attribute, weight_attribute.values.first()
-    )
-
-    attribute_count = Attribute.objects.count()
-
-    category = last_product.category
-    variables = {
-        "filters": {"inCategory": graphene.Node.to_global_id("Category", category.pk)}
-    }
-
-    # when
-    attributes = get_graphql_content(
-        staff_api_client.post_graphql(ATTRIBUTES_FILTER_QUERY, variables)
-    )["data"]["attributes"]["edges"]
-
-    # then
-    assert len(attributes) == attribute_count - 1
-    assert weight_attribute.slug not in {
-        attribute["node"]["slug"] for attribute in attributes
-    }
-
-
-def test_filter_attributes_in_category_not_visible_in_listings_by_app_with_perm(
-    app_api_client, product_list, weight_attribute, permission_manage_products
-):
-    # given
-    app_api_client.app.permissions.add(permission_manage_products)
-
-    product_type = ProductType.objects.create(
-        name="Default Type 2",
-        slug="default-type-2",
-        has_variants=True,
-        is_shipping_required=True,
-    )
-    product_type.product_attributes.add(weight_attribute)
-
-    last_product = product_list[-1]
-    last_product.product_type = product_type
-    last_product.visible_in_listings = False
-    last_product.save(update_fields=["visible_in_listings", "product_type"])
-
-    associate_attribute_values_to_instance(
-        product_list[-1], weight_attribute, weight_attribute.values.first()
-    )
-
-    attribute_count = Attribute.objects.count()
-
-    category = last_product.category
-    variables = {
-        "filters": {"inCategory": graphene.Node.to_global_id("Category", category.pk)}
-    }
-
-    # when
-    attributes = get_graphql_content(
-        app_api_client.post_graphql(ATTRIBUTES_FILTER_QUERY, variables)
-    )["data"]["attributes"]["edges"]
-
-    # then
-    assert len(attributes) == attribute_count
-
-
-def test_filter_attributes_in_category_not_visible_in_listings_by_app_without_perm(
-    app_api_client, product_list, weight_attribute, permission_manage_products
-):
-    # given
-    product_type = ProductType.objects.create(
-        name="Default Type 2",
-        slug="default-type-2",
-        has_variants=True,
-        is_shipping_required=True,
-    )
-    product_type.product_attributes.add(weight_attribute)
-
-    last_product = product_list[-1]
-    last_product.product_type = product_type
-    last_product.visible_in_listings = False
-    last_product.save(update_fields=["visible_in_listings", "product_type"])
-
-    associate_attribute_values_to_instance(
-        product_list[-1], weight_attribute, weight_attribute.values.first()
-    )
-
-    attribute_count = Attribute.objects.count()
-
-    category = last_product.category
-    variables = {
-        "filters": {"inCategory": graphene.Node.to_global_id("Category", category.pk)}
-    }
-
-    # when
-    attributes = get_graphql_content(
-        app_api_client.post_graphql(ATTRIBUTES_FILTER_QUERY, variables)
-    )["data"]["attributes"]["edges"]
-
-    # then
-    assert len(attributes) == attribute_count - 1
-    assert weight_attribute.slug not in {
-        attribute["node"]["slug"] for attribute in attributes
-    }
-
-
-def test_filter_attributes_in_category_not_published_by_customer(
-    user_api_client, product_list, weight_attribute
-):
-    # given
-    product_type = ProductType.objects.create(
-        name="Default Type 2",
-        slug="default-type-2",
-        has_variants=True,
-        is_shipping_required=True,
-    )
-    product_type.product_attributes.add(weight_attribute)
-
-    last_product = product_list[-1]
-    last_product.product_type = product_type
-    last_product.is_published = False
-    last_product.save(update_fields=["is_published", "product_type"])
-
-    associate_attribute_values_to_instance(
-        product_list[-1], weight_attribute, weight_attribute.values.first()
-    )
-
-    attribute_count = Attribute.objects.count()
-
-    category = last_product.category
-    variables = {
-        "filters": {"inCategory": graphene.Node.to_global_id("Category", category.pk)}
-    }
-
-    # when
-    attributes = get_graphql_content(
-        user_api_client.post_graphql(ATTRIBUTES_FILTER_QUERY, variables)
-    )["data"]["attributes"]["edges"]
-
-    # then
-    assert len(attributes) == attribute_count - 1
-    assert weight_attribute.slug not in {
-        attribute["node"]["slug"] for attribute in attributes
-    }
-
-
-def test_filter_attributes_in_category_not_published_by_staff_with_perm(
-    staff_api_client, product_list, weight_attribute, permission_manage_products
-):
-    # given
-    staff_api_client.user.user_permissions.add(permission_manage_products)
-
-    product_type = ProductType.objects.create(
-        name="Default Type 2",
-        slug="default-type-2",
-        has_variants=True,
-        is_shipping_required=True,
-    )
-    product_type.product_attributes.add(weight_attribute)
-
-    last_product = product_list[-1]
-    last_product.product_type = product_type
-    last_product.is_published = False
-    last_product.save(update_fields=["is_published", "product_type"])
-
-    associate_attribute_values_to_instance(
-        product_list[-1], weight_attribute, weight_attribute.values.first()
-    )
-
-    attribute_count = Attribute.objects.count()
-
-    category = last_product.category
-    variables = {
-        "filters": {"inCategory": graphene.Node.to_global_id("Category", category.pk)}
-    }
-
-    # when
-    attributes = get_graphql_content(
-        staff_api_client.post_graphql(ATTRIBUTES_FILTER_QUERY, variables)
-    )["data"]["attributes"]["edges"]
-
-    # then
-    assert len(attributes) == attribute_count
-
-
-def test_filter_attributes_in_category_not_published_by_staff_without_perm(
-    staff_api_client, product_list, weight_attribute, permission_manage_products
-):
-    # given
-    product_type = ProductType.objects.create(
-        name="Default Type 2",
-        slug="default-type-2",
-        has_variants=True,
-        is_shipping_required=True,
-    )
-    product_type.product_attributes.add(weight_attribute)
-
-    last_product = product_list[-1]
-    last_product.product_type = product_type
-    last_product.is_published = False
-    last_product.save(update_fields=["is_published", "product_type"])
-
-    associate_attribute_values_to_instance(
-        product_list[-1], weight_attribute, weight_attribute.values.first()
-    )
-
-    attribute_count = Attribute.objects.count()
-
-    category = last_product.category
-    variables = {
-        "filters": {"inCategory": graphene.Node.to_global_id("Category", category.pk)}
-    }
-
-    # when
-    attributes = get_graphql_content(
-        staff_api_client.post_graphql(ATTRIBUTES_FILTER_QUERY, variables)
-    )["data"]["attributes"]["edges"]
-
-    # then
-    assert len(attributes) == attribute_count - 1
-    assert weight_attribute.slug not in {
-        attribute["node"]["slug"] for attribute in attributes
-    }
-
-
-def test_filter_attributes_in_category_not_published_by_app_with_perm(
-    app_api_client, product_list, weight_attribute, permission_manage_products
-):
-    # given
-    app_api_client.app.permissions.add(permission_manage_products)
-
-    product_type = ProductType.objects.create(
-        name="Default Type 2",
-        slug="default-type-2",
-        has_variants=True,
-        is_shipping_required=True,
-    )
-    product_type.product_attributes.add(weight_attribute)
-
-    last_product = product_list[-1]
-    last_product.product_type = product_type
-    last_product.is_published = False
-    last_product.save(update_fields=["is_published", "product_type"])
-
-    associate_attribute_values_to_instance(
-        product_list[-1], weight_attribute, weight_attribute.values.first()
-    )
-
-    attribute_count = Attribute.objects.count()
-
-    category = last_product.category
-    variables = {
-        "filters": {"inCategory": graphene.Node.to_global_id("Category", category.pk)}
-    }
-
-    # when
-    attributes = get_graphql_content(
-        app_api_client.post_graphql(ATTRIBUTES_FILTER_QUERY, variables)
-    )["data"]["attributes"]["edges"]
-
-    # then
-    assert len(attributes) == attribute_count
-
-
-def test_filter_attributes_in_category_not_published_by_app_without_perm(
-    app_api_client, product_list, weight_attribute, permission_manage_products
-):
-    # given
-    product_type = ProductType.objects.create(
-        name="Default Type 2",
-        slug="default-type-2",
-        has_variants=True,
-        is_shipping_required=True,
-    )
-    product_type.product_attributes.add(weight_attribute)
-
-    last_product = product_list[-1]
-    last_product.product_type = product_type
-    last_product.is_published = False
-    last_product.save(update_fields=["is_published", "product_type"])
-
-    associate_attribute_values_to_instance(
-        product_list[-1], weight_attribute, weight_attribute.values.first()
-    )
-
-    attribute_count = Attribute.objects.count()
-
-    category = last_product.category
-    variables = {
-        "filters": {"inCategory": graphene.Node.to_global_id("Category", category.pk)}
-    }
-
-    # when
-    attributes = get_graphql_content(
-        app_api_client.post_graphql(ATTRIBUTES_FILTER_QUERY, variables)
-    )["data"]["attributes"]["edges"]
-
-    # then
-    assert len(attributes) == attribute_count - 1
-    assert weight_attribute.slug not in {
-        attribute["node"]["slug"] for attribute in attributes
-    }
-
-
-def test_filter_attributes_in_collection_not_visible_in_listings_by_customer(
-    user_api_client, product_list, weight_attribute, collection
-):
-    # given
-    product_type = ProductType.objects.create(
-        name="Default Type 2",
-        slug="default-type-2",
-        has_variants=True,
-        is_shipping_required=True,
-    )
-    product_type.product_attributes.add(weight_attribute)
-
-    last_product = product_list[-1]
-    last_product.product_type = product_type
-    last_product.visible_in_listings = False
-    last_product.save(update_fields=["visible_in_listings", "product_type"])
-
-    for product in product_list:
-        collection.products.add(product)
-
-    associate_attribute_values_to_instance(
-        product_list[-1], weight_attribute, weight_attribute.values.first()
-    )
-
-    attribute_count = Attribute.objects.count()
-
-    variables = {
-        "filters": {
-            "inCollection": graphene.Node.to_global_id("Collection", collection.pk)
-        }
-    }
-
-    # when
-    attributes = get_graphql_content(
-        user_api_client.post_graphql(ATTRIBUTES_FILTER_QUERY, variables)
-    )["data"]["attributes"]["edges"]
-
-    # then
-    assert len(attributes) == attribute_count
-
-
-def test_filter_in_collection_not_published_by_customer(
-    user_api_client, product_list, weight_attribute, collection
-):
-    # given
-    product_type = ProductType.objects.create(
-        name="Default Type 2",
-        slug="default-type-2",
-        has_variants=True,
-        is_shipping_required=True,
-    )
-    product_type.product_attributes.add(weight_attribute)
-
-    last_product = product_list[-1]
-    last_product.product_type = product_type
-    last_product.is_published = False
-    last_product.save(update_fields=["is_published", "product_type"])
-
-    for product in product_list:
-        collection.products.add(product)
-
-    associate_attribute_values_to_instance(
-        product_list[-1], weight_attribute, weight_attribute.values.first()
-    )
-
-    attribute_count = Attribute.objects.count()
-
-    variables = {
-        "filters": {
-            "inCollection": graphene.Node.to_global_id("Collection", collection.pk)
-        }
-    }
-
-    # when
-    attributes = get_graphql_content(
-        user_api_client.post_graphql(ATTRIBUTES_FILTER_QUERY, variables)
-    )["data"]["attributes"]["edges"]
-
-    # then
-    assert len(attributes) == attribute_count - 1
-    assert weight_attribute.slug not in {
-        attribute["node"]["slug"] for attribute in attributes
-    }
-
-
-def test_filter_in_collection_not_published_by_staff_with_perm(
-    staff_api_client,
-    product_list,
-    weight_attribute,
-    permission_manage_products,
-    collection,
-):
-    # given
-    staff_api_client.user.user_permissions.add(permission_manage_products)
-
-    product_type = ProductType.objects.create(
-        name="Default Type 2",
-        slug="default-type-2",
-        has_variants=True,
-        is_shipping_required=True,
-    )
-    product_type.product_attributes.add(weight_attribute)
-
-    last_product = product_list[-1]
-    last_product.product_type = product_type
-    last_product.is_published = False
-    last_product.save(update_fields=["is_published", "product_type"])
-
-    for product in product_list:
-        collection.products.add(product)
-
-    associate_attribute_values_to_instance(
-        product_list[-1], weight_attribute, weight_attribute.values.first()
-    )
-
-    attribute_count = Attribute.objects.count()
-
-    variables = {
-        "filters": {
-            "inCollection": graphene.Node.to_global_id("Collection", collection.pk)
-        }
-    }
-
-    # when
-    attributes = get_graphql_content(
-        staff_api_client.post_graphql(ATTRIBUTES_FILTER_QUERY, variables)
-    )["data"]["attributes"]["edges"]
-
-    # then
-    assert len(attributes) == attribute_count
-
-
-def test_filter_in_collection_not_published_by_staff_without_perm(
-    staff_api_client,
-    product_list,
-    weight_attribute,
-    permission_manage_products,
-    collection,
-):
-    # given
-    product_type = ProductType.objects.create(
-        name="Default Type 2",
-        slug="default-type-2",
-        has_variants=True,
-        is_shipping_required=True,
-    )
-    product_type.product_attributes.add(weight_attribute)
-
-    last_product = product_list[-1]
-    last_product.product_type = product_type
-    last_product.is_published = False
-    last_product.save(update_fields=["is_published", "product_type"])
-
-    for product in product_list:
-        collection.products.add(product)
-
-    associate_attribute_values_to_instance(
-        product_list[-1], weight_attribute, weight_attribute.values.first()
-    )
-
-    attribute_count = Attribute.objects.count()
-
-    variables = {
-        "filters": {
-            "inCollection": graphene.Node.to_global_id("Collection", collection.pk)
-        }
-    }
-
-    # when
-    attributes = get_graphql_content(
-        staff_api_client.post_graphql(ATTRIBUTES_FILTER_QUERY, variables)
-    )["data"]["attributes"]["edges"]
-
-    # then
-    assert len(attributes) == attribute_count - 1
-    assert weight_attribute.slug not in {
-        attribute["node"]["slug"] for attribute in attributes
-    }
-
-
-def test_filter_in_collection_not_published_by_app_with_perm(
-    app_api_client,
-    product_list,
-    weight_attribute,
-    permission_manage_products,
-    collection,
-):
-    # given
-    app_api_client.app.permissions.add(permission_manage_products)
-
-    product_type = ProductType.objects.create(
-        name="Default Type 2",
-        slug="default-type-2",
-        has_variants=True,
-        is_shipping_required=True,
-    )
-    product_type.product_attributes.add(weight_attribute)
-
-    last_product = product_list[-1]
-    last_product.product_type = product_type
-    last_product.is_published = False
-    last_product.save(update_fields=["is_published", "product_type"])
-
-    for product in product_list:
-        collection.products.add(product)
-
-    associate_attribute_values_to_instance(
-        product_list[-1], weight_attribute, weight_attribute.values.first()
-    )
-
-    attribute_count = Attribute.objects.count()
-
-    variables = {
-        "filters": {
-            "inCollection": graphene.Node.to_global_id("Collection", collection.pk)
-        }
-    }
-
-    # when
-    attributes = get_graphql_content(
-        app_api_client.post_graphql(ATTRIBUTES_FILTER_QUERY, variables)
-    )["data"]["attributes"]["edges"]
-
-    # then
-    assert len(attributes) == attribute_count
-
-
-def test_filter_in_collection_not_published_by_app_without_perm(
-    app_api_client,
-    product_list,
-    weight_attribute,
-    permission_manage_products,
-    collection,
-):
-    # given
-    product_type = ProductType.objects.create(
-        name="Default Type 2",
-        slug="default-type-2",
-        has_variants=True,
-        is_shipping_required=True,
-    )
-    product_type.product_attributes.add(weight_attribute)
-
-    last_product = product_list[-1]
-    last_product.product_type = product_type
-    last_product.is_published = False
-    last_product.save(update_fields=["is_published", "product_type"])
-
-    for product in product_list:
-        collection.products.add(product)
-
-    associate_attribute_values_to_instance(
-        product_list[-1], weight_attribute, weight_attribute.values.first()
-    )
-
-    attribute_count = Attribute.objects.count()
-
-    variables = {
-        "filters": {
-            "inCollection": graphene.Node.to_global_id("Collection", collection.pk)
-        }
-    }
-
-    # when
-    attributes = get_graphql_content(
-        app_api_client.post_graphql(ATTRIBUTES_FILTER_QUERY, variables)
-    )["data"]["attributes"]["edges"]
-
-    # then
-    assert len(attributes) == attribute_count - 1
-    assert weight_attribute.slug not in {
-        attribute["node"]["slug"] for attribute in attributes
-    }
-
-
-def test_filter_attributes_by_page_type(
-    staff_api_client,
-    size_page_attribute,
-    product_type_attribute_list,
-    permission_manage_products,
-):
-    # given
-    staff_api_client.user.user_permissions.add(permission_manage_products)
-
-    variables = {"filters": {"type": "PAGE_TYPE"}}
-
-    # when
-    attributes = get_graphql_content(
-        staff_api_client.post_graphql(ATTRIBUTES_FILTER_QUERY, variables)
-    )["data"]["attributes"]["edges"]
-
-    # then
-    assert len(attributes) == 1
-    assert attributes[0]["node"]["slug"] == size_page_attribute.slug
-
-
-def test_filter_attributes_by_product_type(
-    staff_api_client,
-    size_page_attribute,
-    product_type_attribute_list,
-    permission_manage_products,
-):
-    # given
-    staff_api_client.user.user_permissions.add(permission_manage_products)
-
-    variables = {"filters": {"type": "PRODUCT_TYPE"}}
-
-    # when
-    attributes = get_graphql_content(
-        staff_api_client.post_graphql(ATTRIBUTES_FILTER_QUERY, variables)
-    )["data"]["attributes"]["edges"]
-
-    # then
-    assert len(attributes) == len(product_type_attribute_list)
-    assert size_page_attribute.slug not in {
-        attribute["node"]["slug"] for attribute in attributes
-    }
-
-
-ATTRIBUTES_SORT_QUERY = """
-    query($sortBy: AttributeSortingInput) {
-      attributes(first: 10, sortBy: $sortBy) {
-        edges {
-          node {
-            slug
-          }
-        }
-      }
-    }
-"""
-
-
-def test_sort_attributes_by_slug(api_client):
-    Attribute.objects.bulk_create(
-        [
-            Attribute(name="MyAttribute", slug="b", type=AttributeType.PRODUCT_TYPE),
-            Attribute(name="MyAttribute", slug="a", type=AttributeType.PRODUCT_TYPE),
-        ]
-    )
-
-    variables = {"sortBy": {"field": "SLUG", "direction": "ASC"}}
-
-    attributes = get_graphql_content(
-        api_client.post_graphql(ATTRIBUTES_SORT_QUERY, variables)
-    )["data"]["attributes"]["edges"]
-
-    assert len(attributes) == 2
-    assert attributes[0]["node"]["slug"] == "a"
-    assert attributes[1]["node"]["slug"] == "b"
-
-
-def test_sort_attributes_by_default_sorting(api_client):
-    """Don't provide any sorting, this should sort by slug by default."""
-    Attribute.objects.bulk_create(
-        [
-            Attribute(name="A", slug="b", type=AttributeType.PRODUCT_TYPE),
-            Attribute(name="B", slug="a", type=AttributeType.PRODUCT_TYPE),
-        ]
-    )
-
-    attributes = get_graphql_content(
-        api_client.post_graphql(ATTRIBUTES_SORT_QUERY, {})
-    )["data"]["attributes"]["edges"]
-
-    assert len(attributes) == 2
-    assert attributes[0]["node"]["slug"] == "a"
-    assert attributes[1]["node"]["slug"] == "b"
-
-
-@pytest.mark.parametrize("is_variant", (True, False))
-def test_attributes_of_products_are_sorted(
-    staff_api_client, product, color_attribute, is_variant
-):
-    """Ensures the attributes of products and variants are sorted."""
-
-    variant = product.variants.first()
-
-    if is_variant:
-        query = """
-            query($id: ID!) {
-              productVariant(id: $id) {
-                attributes {
-                  attribute {
-                    id
-                  }
-                }
-              }
-            }
-        """
-    else:
-        query = """
-            query($id: ID!) {
-              product(id: $id) {
-                attributes {
-                  attribute {
-                    id
-                  }
-                }
-              }
-            }
-        """
-
-    # Create a dummy attribute with a higher ID
-    # This will allow us to make sure it is always the last attribute
-    # when sorted by ID. Thus, we are sure the query is actually passing the test.
-    other_attribute = Attribute.objects.create(
-        name="Other", slug="other", type=AttributeType.PRODUCT_TYPE
-    )
-
-    # Add the attribute to the product type
-    if is_variant:
-        product.product_type.variant_attributes.set([color_attribute, other_attribute])
-    else:
-        product.product_type.product_attributes.set([color_attribute, other_attribute])
-
-    # Retrieve the M2M object for the attribute vs the product type
-    if is_variant:
-        m2m_rel_other_attr = other_attribute.attributevariant.last()
-    else:
-        m2m_rel_other_attr = other_attribute.attributeproduct.last()
-
-    # Push the last attribute to the top and let the others to None
-    m2m_rel_other_attr.sort_order = 0
-    m2m_rel_other_attr.save(update_fields=["sort_order"])
-
-    # Assign attributes to the product
-    node = variant if is_variant else product  # type: Union[Product, ProductVariant]
-    node.attributesrelated.clear()
-    associate_attribute_values_to_instance(
-        node, color_attribute, color_attribute.values.first()
-    )
-
-    # Sort the database attributes by their sort order and ID (when None)
-    expected_order = [other_attribute.pk, color_attribute.pk]
-
-    # Make the node ID
-    if is_variant:
-        node_id = graphene.Node.to_global_id("ProductVariant", variant.pk)
-    else:
-        node_id = graphene.Node.to_global_id("Product", product.pk)
-
-    # Retrieve the attributes
-    data = get_graphql_content(staff_api_client.post_graphql(query, {"id": node_id}))[
-        "data"
-    ]
-    attributes = data["productVariant" if is_variant else "product"]["attributes"]
-    actual_order = [
-        int(graphene.Node.from_global_id(attr["attribute"]["id"])[1])
-        for attr in attributes
-    ]
-
-    # Compare the received data against our expectations
-    assert actual_order == expected_order
