+++ saleor/graphql/product/types/__init__.py
@@ -1,5 +1,4 @@
 # flake8: noqa
-from .attributes import Attribute, AttributeValue, SelectedAttribute
 from .digital_contents import DigitalContent, DigitalContentUrl
 from .products import (
     Category,
